GH-1474 create AddressTimestamp object; refactor
